+++ source/ComponentManager.js
@@ -2,12 +2,15 @@ const fs           = require('fs');
 const path         = require('path');
 const vscode       = require('vscode');
 const fileContents = require('./FileContents');
+const Logger       = require('./Logger');
 const { SanitizeFileName, GetWorkspacePath } = require('./CommonUtils');
 
 let createComponentDisposable;
 
 /**
- * @param {*} context 
+ * Registers the 'createComponent' command in the extension.
+ * 
+ * @param {vscode.ExtensionContext} context The extension context provided by VSCode.
  */
 function CreateComponentCommand(context)
 {
@@ -22,9 +25,11 @@ function CreateComponentCommand(context)
 class Component
 {
     /**
-     * @param {string?} name
-     * @param {boolean} mocked
-     * @param {boolean} tested
+     * Creates a new Component instance.
+     * 
+     * @param {string?} name   The name of the component.
+     * @param {boolean} mocked Indicates whether the component is mocked.
+     * @param {boolean} tested Indicates whether the component is tested.
      */
     constructor(name, mocked, tested) 
     {
@@ -35,12 +40,16 @@ class Component
 }
 
 /**
+ * Prompts the user to select properties for a component.
  * 
- * @param {Component} component 
- * @returns 
+ * @param {Component} component The component for which to select properties.
+ * 
+ * @returns {Promise<Component|undefined>} The updated component, or undefined if the selection was cancelled.
  */
 async function SelectComponentProperties(component)
 {
+    Logger.Info('Attempting to create a component');
+
     let componentName = await vscode.window.showInputBox({ prompt: 'Enter the name of the new component' });
     if (componentName !== undefined)
     {
@@ -48,6 +57,8 @@ async function SelectComponentProperties(component)
     }
     else
     {
+        Logger.Warning('Unable to create a component: component name is required.');
+        vscode.window.showWarningMessage('Component name is required.');
         return undefined;
     }
 
@@ -73,7 +84,11 @@ async function SelectComponentProperties(component)
     });
 
     // if no selection, do not do anything
-    if (!selectedProperties) return undefined;
+    if (!selectedProperties)
+    {
+        Logger.Error('No component properties were selected.');
+        return undefined;
+    }
 
     // Reset properties to false
     properties.forEach(prop => 
@@ -91,12 +106,12 @@ async function SelectComponentProperties(component)
 }
 
 /**
+ * Composes a list of files to be created for a component.
  * 
+ * @param {Component} component        The component to compose files for.
+ * @param {string}    componentDirPath The directory path where the component files will be located.
  * 
- * @param {Component} component 
- * @param {string}    componentDirPath 
- * 
- * @returns 
+ * @returns {Array<{path: string, content: string}>} An array of file objects with path and content properties.
  */
 function ComposeComponentFiles(component, componentDirPath) 
 {
@@ -127,11 +142,12 @@ function ComposeComponentFiles(component, componentDirPath)
 }
 
 /**
- * Creates a directory for the component inside the "components" directory
+ * Prepares a directory for the component inside the "components" directory.
  * 
- * @param {Component} component 
+ * @param {Component} component The component for which to prepare the directory.
  * 
- * @returns undefined if the component folder already exists
+ * @returns {Promise<string|undefined>}
+ * The path to the component directory, or undefined if the folder already exists.
  */
 async function PrepareComponentDirectory(component)
 {
@@ -139,6 +155,7 @@ async function PrepareComponentDirectory(component)
 
     if (fs.existsSync(componentDirPath))
     {
+        Logger.Error(`Component "${component.name}" already exists.`);
         vscode.window.showWarningMessage(`Component "${component.name}" already exists.`);
         return undefined;
     }
@@ -149,11 +166,12 @@ async function PrepareComponentDirectory(component)
 }
 
 /**
- * Adds the newly created component to the main CMakeLists.txt
+ * Adds the newly created component to the main CMakeLists.txt.
  * 
- * @param {Component} component 
+ * @param {Component} component The component to register in the CMakeLists.txt.
  * 
- * @returns undefined if CMakeLists.txt does not exist.
+ * @returns {Promise<void|undefined>}
+ * A promise that resolves when the operation is complete, or undefined if CMakeLists.txt does not exist.
  */
 async function RegisterComponentToMainCmake(component)
 {
@@ -161,7 +179,8 @@ async function RegisterComponentToMainCmake(component)
     let rootCmakeFilePath = path.join(GetWorkspacePath(), 'CMakeLists.txt');
     if (fs.existsSync(rootCmakeFilePath) === false)
     {
-        vscode.window.showWarningMessage(`Root CMakeLists.txt not found.`);
+        Logger.Error(`Root CMakeLists.txt not found.`);
+        vscode.window.showErrorMessage(`Root CMakeLists.txt not found.`);
         return undefined;
     }
 
@@ -198,9 +217,11 @@ async function RegisterComponentToMainCmake(component)
 }
 
 /**
- * Creates a new component by asking the user about component name and properties
+ * Handles the creation of a new component.
  * 
- * @returns undefined if component creation was cancelled or it already existed
+ * @returns {Promise<void|undefined>}
+ * A promise that resolves when the component is created, or undefined if the creation
+ * was cancelled or the component already existed.
  */
 async function createNewComponent()
 {
@@ -221,7 +242,7 @@ async function createNewComponent()
     let component = new Component(undefined, false, false);
 
     await SelectComponentProperties(component);
-    if (component === undefined) return undefined;
+    if (component.name === undefined) return undefined;
 
     let componentDirPath = await PrepareComponentDirectory(component);
     if (componentDirPath === undefined) return undefined;
@@ -230,6 +251,8 @@ async function createNewComponent()
     files.forEach(file => fs.writeFileSync(file.path, file.content));
     
     await RegisterComponentToMainCmake(component);
+
+    Logger.Info(`Component ${component.name} created.`);
 }
 
 module.exports = CreateComponentCommand;